Ensure current-{input,output}-port are properly restored on exception.
Modify with-input-from-{file,pipe} and with-output-to-{file,pipe} to use
fluid-let on ##sys#standard-{input,output} to ensure they are restored
after an exception.  The ports were restored only if the exception bubbled
up to the REPL; if intercepted prior to that, the REPL would then read
further commands from (or write output to) that file or pipe.

Signed-off-by: felix <felix@call-with-current-continuation.org>
(cherry picked from commit 5ae71ecd963948d0d1c9eeea5c47c32ecaac963a)