Mark all unused procedures (or replaced by compiler-macro) in irregex to get rid of compile-time warnings
Signed-off-by: felix <felix@call-with-current-continuation.org>
(cherry picked from commit 84e997b979393600323c19d2149c39c88b4ea71e)

Backporting:

irregex-match-chunk&index-from-tag-set! is not present because
c5a5d250c38 (Laurikari's algorithm) has not been applied, so the
unused declaration gives a conflict.  We add the definition anyway
as it is harmless and, if the patch is later applied, we need to
keep the unused declaration.

Conflicts:

	irregex.scm