Make the "-module" option take the module name as an argument
Allows specifying the module name at the command line, aliases this
option as "-m", and adds a "-main-module" option (aliased as "-M") that
provides the old behaviour (a default module name of "main").

Also removes the long-since-deprecated "-scrutinize" option and fixes a
missing import of c.c.support in chicken.scm (for `quit-compiling`, and
possibly other procedures).