Do not override ##sys#quit-hook in csi
This is not needed; the "repl" procedure already overrides it anyway.

The old behaviour seems to have been that this (incorrectly!) created
a "local" ##sys#quit-hook in csi which ",q" would invoke, ignoring the
global one.  However, it's fine to simply call the default quit hook.

Instead of directly calling ##sys#quit-hook (which has a different
signature than the aforementioned "local" quit-hook), we now call the
official API: "quit".  This has an optional argument, so it won't
bomb out like ",q" did before this patch.

Signed-off-by: Evan Hanson <evhan@foldling.org>