Fix "identify.sh" for release and snapshot builds.
This is a bit of a hack: check whether the manual-html directory
exists and assume it must be a distribution build, then exit to
avoid a rebuild which requires a pre-existing chicken.

Changeset 4e1f7272a3536dd93577c71f0a8bad16dab2a308 (for "allowing
custom builds") caused this breakage.