Avoid adding duplicate file descriptors when populating C_fdset_set
Previously, when populating the C_fdset_set with pollfd entries before
unblocking threads for i/o, an entry was added for each thread waiting
on a given file descriptor, meaning multiple entries might be created
for the same fd. Later code assumes that the C_fdset_set is a true set
of fds, causing an assertion to fail. This patch changes `create-fdset`
to ensure that each file descriptor appears once in the C_fdset_set,
with the union of its input and output flags.