Fix heap allocation size calculations in toplevel.
The toplevel is responsible for decoding literals and storing them in
the heap.  When the heap is not big enough to store one toplevel's
literals, the toplevel will call C_rereclaim2 with the desired extra
size for the heap.

C_rereclaim2 will calculate the size by adding it to the current heap
size, but because the heap is split in two parts, it needs to multiply
the demanded size with two.

This change also adds the current stack size to this, because when the
GC is performed, it will also copy what's in the nursery to the heap. So
worst case, the heap size needs to be (current heap size + demanded
memory + stack size).

We add regression tests for this as well.  This fixes #1221.

Signed-off-by: Evan Hanson <evhan@foldling.org>