Copy thread "state vector" to a fresh argvector before invoking the trampoline function.
Upon resumption of a thread from a "state vector" (rather than a
closure/continuation), the trampoline function was passed a pointer
right into the state vector, which doubled as an argument vector.

Now that argvectors may be re-used, their slots can be trashed any
time in the call history, depending on the exact situation.  These
argvector slots will eventually contain temporary data which becomes
invalid over time.  When the original state vector is then copied
during garbage collection, the mutated slot is also copied.  When the
Cheney algorithm traverses over this slot, it will break down because
the data it points to is now invalid.

This hopefully fixes the "random crashes" we've been seeing on
Salmonella the last few weeks.

Signed-off-by: felix <felix@call-with-current-continuation.org>