Fix equal? comparison of closures.
Even though you're not supposed to compare closures (the spec says it's
undefined), CHICKEN supports this to a degree.  Two closures are
compared in such a way that the slots must be identical.  In other
words, the function address and all the variables that are closed over
are eq? to one another on both sides.

The current implementation had an off-by-one which caused it to omit the
comparison of the last slot, which holds either the last argument or the
lambda-info in case we've compiled with lambda info.  This caused the
equal? procedure to behave differently depending on whether the code was
compiled with the -no-lambda-info option or without it.

Thanks to Mario Goulart for finding this in ticket #1041.

Signed-off-by: Mario Domenech Goulart <mario.goulart@gmail.com>