Add a basic form of source-level debugging to compiled Scheme code
The compiler has been extended to decorate code with "debug-event"
forms ("##core#debug-event"). In the code-generation phase these
emit calls to a hook-function ("C_debugger") that is used to
submit debug events to a connected TCP server. Each event has
an associated record of debug-information (line-numbers, etc.)

An extension to the run-time system implements the C_debugger hook
and provides a simple TCP-based connection to a TCP server on
port 9999, sending events, when encountered and reading response
"actions" from the server. The protocol is text- and s-expression
based.

The new debug-level 3 enables the mechanism. If "-d3" is given,
the "debugger-client" unit is invoked, which checks the environment
variable "CHICKEN_DEBUGGER" (that should hold a string of the form
"<servername>:<port>"), tried to connect and, if the connection
can be established, suspends the program, waiting for responses.

A simple Tcl/Tk based debugger "front-end" is provided. It is called
"feathers" and requires Tcl/Tk 8.5.

Basic usage is explained in the new "Debugging" chapter in the
manual. It cointains just the basics, though. I have tested this
on Linux, Windows and Mac. As the version of Tcl/Tk on Mac is
very buggy, I was only able to run it with an X11-based version
(requires XQuartz). It seems to run ok on Windows (using ActiveState
Tcl/Tk), but I have only done very basic tests on that platform.

The front-end currently allows single-stepping, inspecting arguments
and global variables, setting break-points on source lines (only
for those, for which line-number information could be recovered)
and global variables.

The build has been extended to produce and install executable scripts
("feathers" and "feathers.bat"), and the "debugger-client" unit
has been added.

Note that the calls to C_debugger have a performance impact.
"C_trace" is called implicitly in the case of a "procedure call"
event to avoid an additional library function call.
C_debugger is called for the following events: procedure
entry, procedure call, global-variable assignment, GC, error and
termination.

Signed-off-by: Evan Hanson <evhan@foldling.org>