Rehaul library loading
Import libraries now include a library identifier that is consulted in
order to determine a module's presence. When loaded, a module provides
this identifier via `##sys#provide` from the eval unit.

When shared libraries are loaded with `##sys#require`, either via
`import` or some similar method, an attempt is made to enter a toplevel
corresponding to the required library's name before falling back to the
library's default, non-unit toplevel.

Units specified by the "-uses" flag are treated identically to those
declared used in-source.

The "-R" flags to csc(1) and csi(1) now use `import` to load the
requested libraries.