Fix nonlocal exit handling in parameterize
When a parameter's "guard" procedure raises an exception, the values
should be consistent.  Before, the values were assigned one by one, and
if the exception got raised halfway through, the already-assigned
parameters would not be restored to their original values.

Add a clarifying comment as to why we must alias all the parameters:
they don't need to be identifiers but can be complex expressions
evaluating to parameters.  This is not immediately obvious from the
code.  It's only clear when you remember that parameterize can be used
like this when reading the code.

Signed-off-by: Evan Hanson <evhan@foldling.org>