Port another B/Z performance fix from numbers egg.
This is actually the *proper* fix for the pathological case that
revision b0711c4ca7f8e92dd18807a1c553e6ce649f1380 tried to fix: Because
X is divided into DIV_LIMIT (BZ_DIFF_THRESHOLD) sized pieces, this only
makes sense if Y is bigger than those pieces.

Finally, the revision in which I tried to fix it also changed the name
of the threshold parameter, but it didn't fix the #ifdef guard around
it, which means it wasn't overrideable by the user like it should be.
This changes back the name to the original, thereby fixing that problem
as well.