Replace strndup() with strdup() for compat reasons
This is the only thing preventing successful build for MingW on Vista.

Also fix the default port, if only a plain hostname was supplied in the
CHICKEN_DEBUGGER environment variable.  Before, it would have been
reading an uninitialized variable and always copying the full string as
a host name due to an unintended cast to unsigned because of a bad stop
condition in the loop.

Signed-off-by: Evan Hanson <evhan@foldling.org>