Fix scrutiny out/expected comparison in tests on Mingw w/o MSYS
It turns out that "fc", Microsoft's pathetic excuse for a file
comparison tool, only reads in a fixed number of lines when comparing
files (100 by default).  If that number is exceeded, it just stops and
reports there were differences, even though there are only newline
differences, even when whitespace is ignored.

So we just increase the buffer size and hope for the best.

Signed-off-by: Evan Hanson <evhan@foldling.org>