Load lolevel in compiler so that foldable bindings are present
This ensures that all identifiers that are subject to constant folding
are bound to procedures in the compiler so that `constant-form-eval`
will work correctly.

Note that the data-structures, extras, and srfi-4 units also contain
foldable bindings, but these are already used by the compiler so they
don't need to be added.

Signed-off-by: Peter Bex <peter@more-magic.net>