Fix excessive temporary generation
In deeply nested if structures, we would generate a set of temporaries
for each arm of the if, while the code would re-use temporaries.

This means if each arm uses N temporaries, at a nesting depth of M we
would generate N*2^M temporaries, instead of just N.

This is especially clear in the generated code for core.scm; there is
one C function with over 200 temporaries which uses less than 100!

Signed-off-by: Evan Hanson <evhan@foldling.org>