Include argvector size in C_demand() calculations.
When generating code, we check all the callees and their argument sizes.
The largest of these will be added to the C_demand() call to ensure we
can allocate the argvector to call it.  If this is less than the current
argument count, we don't add it because we won't need to allocate.

Signed-off-by: Evan Hanson <evhan@foldling.org>