Don't shortcut pure calls in the optimiser
When we notice a node represents a call to a pure (side effect-free)
procedure, we attempt to drop the call completely, if the result is
unused.  However, if the result _is_ used, we shouldn't just give up.
In fact, if the callee is of type explicit-rest, we *must* finish the
optimisation by tweaking the caller.

This fixes a bug reported by Joerg Wittenberger which got triggered
rather easily when adding profiler instrumentation to procedures,
because those always use rest args.

Signed-off-by: Evan Hanson <evhan@foldling.org>