Fix fixnum literal decoding on ARM64 (#1280)
In literal_decode, we relied on sign extension of characters when
putting them into a C_word.

On ARM64, characters are unsigned by default, which causes fixnum
literals (but only if embedded in other objects) to be decoded
incorrectly.

The fix is simple: explicitly cast the first value to a signed character.

Many thanks to Alexander Shendi for reporting the bug and helping to
come up with and testing this fix.

Signed-off-by: Evan Hanson <evhan@foldling.org>