Change how unbound variables are checked for
The previous method checks if the object 'x' we want to check is equal
to the value contained in the first slot of a symbol that's supposed to
always be undefined.

This patch replaces the code above with the use of the C_boundp and
C_unboundvaluep macros with no functional effect.

Signed-off-by: Evan Hanson <evhan@foldling.org>
Signed-off-by: Peter Bex <peter@more-magic.net>