Make `sleep` suspend thread if scheduler is loaded, process otherwise
Makes `sleep` change its behaviour based on whether the scheduler is
loaded: when it is, the current thread is suspended (allowing other
threads to continue executing), and when it isn't the whole process is
suspended with sleep(3). Also adds a `process-sleep` procedure to the
posix unit as a way to sleep the process unconditionally.

Signed-off-by: Peter Bex <peter@more-magic.net>