Fix several hygiene issues in DSSSL expansions (#806).
The expansion of DSSSL lambda list would take "optional", "let*" and
"let-optionals*" from the current syntactic environment, which might
not have it or a different definition, depending on the user's import
statements.

Now we explicitly take them from the default macro environment or
the chicken macro environment, depending on the macro to use.

Signed-off-by: Evan Hanson <evhan@foldling.org>