Do not track set! to known-to-be-immediate values.
This extends the core language by adding an optional second param to the
internal core version of set!, which gets filled in by the scrutinizer
whenever the set! value is determined to always be immediate.

In tight loops where set! is called very often it might make a
difference.  On our benchmarks it doesn't make a dent in the results,
though.

Signed-off-by: Evan Hanson <evhan@foldling.org>