Fix buffer overflow in posix execvp/execve wrapper
This fixes bug #1308 found by wasamasa.  It turns out that we don't
check the number of arguments or the number of env entries before trying
to write them to the target string.

Instead of checking the argument count, this patch replaces the static
buffer with a dynamically allocated string and relies on errno being
set to E2BIG if the argument vector is too large.

Furthermore, this merges the process-execute and process-spawn
code from Windows and Unix some more to use more common code.
This should make it easier to tweak this code in the future.

This new version also fixes a memory leak which would be
triggered when the arg or env list contained non-string objects
or embedded NULs, or when the exec itself would fail.

Most C code in these procedures was rewritten to Scheme.

Signed-off-by: Peter Bex <peter@more-magic.net>
Signed-off-by: Evan Hanson <evhan@foldling.org>