Use "~a" formatter for procedure names in scrutiny output
In some `list-ref` and `vector-ref` calls for which an
index-out-of-range scrutiny message is printed, the procedure name is a
string rather than a symbol. We should find where this difference is
introduced and fix it, but until then we need to use the "~a" output
formatter so these procedure names are printed without surrounding
quotes.