Do not warn for out of range indices into possibly smashed list types
When a list is smashed, usually ends up as (or pair null).  If then we
cons something onto it, it's seen as a list of length 1 or possibly 2.
We should *not* give a warning on (list-ref 3 that-list), because it may
originally have been a list of a greater length.  We don't know that, so
we should avoid warning for anything that's not absolutely sure to be a
proper list.

Luckily, if it's typed as a proper list, that's presumably safe.  That's
because a list with smashed components should end in just "pair", due to
possible mutation by set-cdr!, which means its type is not that of a
proper list.

We still always warn when list-ref takes a negative index, because
that's never ever valid, regardless of the argument list type.

We still always preserve types when using list-ref, even on a list with
smashed components, as long as the list is known to contain the index.

Signed-off-by: Evan Hanson <evhan@foldling.org>