Change bignum representation to use a custom type.
Instead of using a generic record type with a "bignum_type_tag" symbol
slot, we use a custom object containing only a pointer to the actual
bignum data.

This saves at least one slot when pre-allocating for a call to an
arithmetic operator.  In case of the generic operators, this can save up
to four slots (maximum size is allotted for a complex number consisting
of two ratnums, each containing two bignums; so four bignums in total).

Less allocation results in slower nursery growth, which in turn means
minor garbage collections will happen less frequently.

The change in object type also means we need to do one less check in
the bignum? predicate.

Signed-off-by: felix <felix@call-with-current-continuation.org>