Change ratnum and cplxnum representation to also use custom types.
Same change as bignums.  This eats up two more reserved type tags,
resulting in more pre-allocation savings, about as much as with the
bignum change.

The performance gains on code that *doesn't* use these numeric types
are minimal, but it really cleans up the code; the numeric ops are now
much less branchy, and it changes 2 checks to 1 check for every case
where a ratnum or a cplxnum is involved.  In total, this is a net
removal of 100 lines of code.

While at it, this also replaces direct C_block_item(x, n) calls with
more opaque/abstract C_u_i_ratnum_{num,denom} and C_u_i_cplxnum_{real,imag}
helper macros.  This would make it easier to change the representation in
the future.  Currently only the eqv? implementation directly accesses the
slots to keep the code simpler.

Signed-off-by: felix <felix@call-with-current-continuation.org>