Add `include-relative` form to "chicken" module
Adds a new include form that searches for files relative to the
including file rather than the CWD. In all other ways it behaves like
the normal `include`.

Because `load-relative` and `include-relative` are so similar, this
change also moves some pathname handling code into a group of shared
procedures in eval.scm and uses them for both.

Also slightly simplifies the filename handling code in `load` and fixes
a segfault in the following program caused by the use of
`##sys#current-load-path` as a string when it's false:

  (load (open-input-string "(load-relative \"foo.scm\")"))

Signed-off-by: Peter Bex <peter@more-magic.net>