Fix symbol GC: add wep lookup after fptr chasing
Sometimes, with symbol GC enabled, a major GC might "drop" symbols which
were still being referenced, resulting in weird errors like (eq? x 'foo)
returning #f even if x holds the symbol 'foo.

If, during marking in major GC, we encounter the bucket before we
encounter the symbol, the bucket still refers to the symbol in its
original location (the fromspace).  This pointer is added to the weak
table with a counter of 0.

Then, the symbol itself is scanned, and the item is found in the weak
table, the counter is updated and the symbol is moved to the heap.  The
header at the symbol's original location in the fromspace becomes a
forwarding pointer.

Then, when we encounter a _second_ reference to the symbol, it still
refers to the symbol's pointer in the fromspace, but the header will be
a forwarding pointer, so it won't match the symbol type (which we look
for right at the start of the mark function).  This means the code code
to update the weak entry's count won't be triggered.

Instead, we should chase the forwarded pointer and *then* check if it's
a symbol.  If it is, look up the *original* location's pointer in the
weak table.

Note: We don't need to look up the new location, because that can only
be the case if the symbol was marked before we encountered the bucket,
in which case it will already saturate the pointer immediately upon
insertion of the weak table entry.

Note 2: Before a reallocing GC, we reset the weak table and we never
consult it during the reallocing GC, so all symbols will be copied.  A
minor GC also doesn't handle symbols specially, so they'll be copied
there too.

This fixes #1173

Signed-off-by: felix <felix@call-with-current-continuation.org>