Remove C_BUCKET_TYPE
The original GC code detected C_BUCKET_TYPE as a special case, for which
it would insert the bucket's symbol into the weak symbol hash table.

Now that we treat hsah table buckets *almost* normally, there's only one
special case left, which also works if we use "weak pair" as a type.

There is one caveat, namely that anything which checks for "pair" by
directly matching the header tag against C_PAIR_TAG will not accept a
weak pair.  This is not as bad as it sounds, since the buckets are in an
internal hash table that's not directly accessible to the user anyway.
The only place we need a special case for it is in dump_heap_state_2().

Signed-off-by: Evan Hanson <evhan@foldling.org>