Detect and signal error on stack overflow in `apply`
When a demand for nursery space is unsatisfied even after popping the
stack and reinvoking the caller, going back to C_save_and_reclaim to try
to reclaim space *again* just leads to an infinite loop. Since there's
not much else we can do in this situation, it's better to signal the
problem as a Scheme-level error.

To do this, we can just remember that there was a stack demand when
jumping into the GC from C_apply or C_apply_values and, if the same
demand fails when we bounce back and try again, we barf.

The one tricky thing to this is that we must forget the remembered
demand if the jump to the GC is redirected to the interrupt handler.
Otherwise, the next call to C_apply/C_apply_values would only get one
chance to demand nursery space.