(by Peter Bex) Simplify and improve reclaimability of symbol GC
Instead of using a secondary symbol table ("weak table") for tracking
unreferenced symbols, we change the symbol's buckets to either hold the
symbol strongly or weakly.  This allows us to rely on the GC's normal
function to copy or discard unused symbols.

On minor GC, we *always* copy the symbols, ignoring its bucket's
weak/strong status.  This is for performance reasons.

After a major or reallocating GC, we traverse the symbol table and
update the pointers in weak buckets: if the new location of the symbol
is in the tospace/new heap, it must be live, so we keep it and update
the bucket's pointer.  Otherwise, we can drop the bucket.

This requires one change to how symbols are managed: when a bucket is
allocated, it is weak by default.  When a symbol becomes globally bound
or gains a plist, we change the bucket type from weak to strong.  If the
plist is emptied, the bucket type is changed to weak again (but only if
it isn't also globally bound).

Currently we don't support unbinding a symbol, but if we ever do we'll
need to call unpersist on the symbol as well.

There are several advantages to this:
1) It is simpler than the tricky weak table code (see also #1173).
2) Now, all symbols can be collected instead of an upper limit of 997.
3) It is much faster when a lot of discardable symbols are generated.
4) Memory usage is no longer unbounded when many symbols are generated.

Points 3 and 4 are related to point 2.  See the knucleotide benchmark
for an extreme example of this: it uses about half as much memory and
is twice as fast under the new implementation.

Signed-off-by: felix <felix@call-with-current-continuation.org>