(by Peter Bex) Make weak symbol GC the default.
Because symbol GC is now nothing special, we can do this.  The overhead
of the symbol table fixup in update_symbol_tables() is neglible, and in
many cases offset by the gains of not needlessly copying symbol data.

This removes the -:w option, and the SYMBOLGC compile-time option no
longer does anything.

Signed-off-by: felix <felix@call-with-current-continuation.org>