Remove a few more semi-unused fudge factors.
Fudges 8 (words per double), 9 (latency), 10 (clocks per sec),
23 (startup time), 29 (trace buffer size) were used nowhere.

Remove ##sys#apply-limit, as it is used nowhere and doesn't make much
sense (its value is just a crude guess anyway).  This was the only use
of fudge 34.

Fudge 36 (toggle debug mode) doesn't make much sense either.  If we want
to set it it is more logical to just set it to the desired value.  The
one test that used it seemed to use it for no apparent reason (??)

Signed-off-by: Evan Hanson <evhan@foldling.org>