Add helper functions for remaining fudge factors.
These are all used directly inline in the one place they were used.

For fudge factor 13 (debug mode), a ##sys#debug-mode? is added, to make
it more readable.  A types.db entry is provided to ensure debug mode
checks don't cause a performance regression (there was a built-in
rewrite for ##sys#fudge).

Alternatively, some of the internal variables could be exposed to the
outside world (like fake_tty, finalizer_count etc), but this would also
make them settable, which is something we might not want.  And by not
having them be static, performance could also be adversely affected in
case of the variables used in the GC.

Signed-off-by: Evan Hanson <evhan@foldling.org>