Improve read/write invariance of keywords (#1332).
Keywords are now treated more like symbols are: when they are written,
we check for readability, which means they'll be pipe-delimited if they
contain "special" characters.

The reader now also uses the same "extended token" reader for keywords
using the "portable" representation as the style-specific reader, so
that it's possible to enter keywords containing "special" characters
when using this style.

We now also support empty keywords, which can be entered through the
"portable" syntax using quotation, i.e., as #:||.  Currently, ||: and
:|| are not recognised as keywords, but as a symbol of one character,
the colon.  Ie, it's treated the same as : by itself.

Signed-off-by: Evan Hanson <evhan@foldling.org>