Remove run-time/compile-time diffs in parameterize
We bind the "saveds" once more to their own values.  This ensures that
any call/cc in the parameterize form itself will not affect the
saving/restoring of the parameters due to compilation strategy.

Strictly speaking, I think this is not required by SRFI-39, but this way
we get consistent behaviour in the tests, compiled or interpreted.

Second part of fix for #1336

Signed-off-by: Evan Hanson <evhan@foldling.org>