Fix crashes in callbacks after GC (#1337).
This was caused by keeping argvector in temp stack, similar to 9eed2742.
Instead, we also copy the argvector into the stack before re-invoking
the trampoline when in a callback context, just like a regular CPS
context.

Also, fix an off-by-one in the number of arguments which
allocate_vector_2 accepts; simply passing the original argcount doesn't
allow for the extra 'mode' argument which gets added onto the front of
the stack.

Signed-off-by: Christian Kellermann <ckeen@pestilenz.org>