Make "time" command detection more robust in tests
This is a better way of checking for a "time" command: we just call it
and check for an exit status of 127 (i.e. command not found).

Note that `time' must be called through a variable during this check.
Otherwise, the Cygwin shell will use its builtin for the test, but not
later in the script when it's called as "$time", causing an error.