Change the way LET bodies are macro-expanded.
A macro might expand into a define.  That means we need to keep
expanding the body and restart the main expansion process when we
encounter a define.  Instead of returning the original expressions
when wrapping up, we should return the macro-expanded expressions,
because macros should be called exactly once to be safe in the
presence of side-effects.

We now also treat ##core#begin as a reason to restart the expansion
process, because nested begins can contain definitions as well.  The
expansion will recursively eliminate those nested begins.

There is some special treatment for ##core#module and "import",
because those do all kinds of nasty side-effecting things which ensure
we can't simply expand the body in one go.  Import is one of those
aforementioned side-effecting macros, and the core module form is also
side-effecting in a way: we can't refer to the module until it has
been processed by the compiler.

There could be other macros and special forms that are allowed in let
bodies but need special processing.  This situation needs to be
addressed properly and fixed in general, but for now we can fix them
by adding more special cases.  Note that this is not a newly
introduced problem: there have always been such issues, but due to
the obscure workings of ##sys#canonicalize-body they would only
surface under very specific conditions.

Signed-off-by: felix <felix@call-with-current-continuation.org>