Fix running of tests when not installed (#1347)
When the tests were being run, and an extension was required for
syntax, the compiler would not be able to load the extension or even
its import library, because it would search the target installation
directory, which either contained an old, outdated, version of the
library (whoops!), or it wouldn't exist at all.

To fix this we simply export CHICKEN_REPOSITORY, always.

Static linking was also busted for much the same reason: it would link
directly against /path/to/chicken/lib/libchicken.a.  This is not
necessary, because the path would be put after -static.  Instead, we
rely on ld(1)'s behaviour that any -l flags after the -static flag is
interpreted as referring to a static archive, so we just use -lchicken
in all cases, regardless of requiring static or dynamic linking.

This should also fix static cross-compilation when the target library
name of libchicken has been customised.

Signed-off-by: Evan Hanson <evhan@foldling.org>