Avoid allocating scratchspace when reading fixnums.
The code that handles reading digit streams into an integer is written
to be as straightforward as possible.  Currently this means it always
allocates a bignum, even if the number is small enough to fit a
fixnum.  That way, we don't need to distinguish between these cases
in the actual stream converter (str_to_bignum).

We can improve on this a little bit by pre-allocating enough for a
small bignum in the nursery, much like the C_s_a_u_i_integer_plus and
friends do.  This avoids malloc()ing a scratchspace just for the one
small number, which will be free()d again as soon as the stack is
full, resulting in a lot of needless thrashing.

Signed-off-by: Evan Hanson <evhan@foldling.org>