Remove inconsistent typecast to C_restart_trampoline
The function argument "trampoline" is already of type void *, which is
the type of C_restart_trampoline, to which we assign it.  Casting to
C_proc results in a compiler warning on Sun Studio.

This was introduced in 67a3d9c4, when we got rid of C_HACKED_APPLY and
the TRAMPOLINE typedef.

Signed-off-by: Evan Hanson <evhan@foldling.org>