Fix a bug in scan-toplevel-assignments walk routine
Commit [ac8f2dadd] introduced a bug in the node walking routine. We end
up evaluating only the first two subexpression for some kind of nodes,
this means the else branch for if/cond nodes is never walked and neither
is most of ##core#switch since we stop the walk at the first constant
node.

Signed-off-by: Peter Bex <peter@more-magic.net>
Signed-off-by: Evan Hanson <evhan@foldling.org>