Do not undo macro renaming when renaming twice.
The "lookup" procedure will try to look up the renamed identifier in
the syntax env and if that fails, it checks whether the looked-up
identifier is an alias of another identifier, in which case it
returns the original identifier.

This fallback behaviour causes problems when we rename an identifier
twice.  In normal use, that doesn't tend to happen (but it could).
When using IR macros, this happens all the time, because the input
form is completely renamed first, and then the output form is
reverse-renamed (renaming any "plain" identifiers in the process).
This behaviour will "collapse" any plain identifiers, so that, on the
reverse rename, they are mapped to the same identifier as those from
the input form.

This fixes ticket #1362, as found by Megane.

Signed-off-by: Evan Hanson <evhan@foldling.org>