Remove unnecessary double alias for already aliased renamed identifiers.
This was an overly complicated solution for #518, which was only
necessary because of the bogus "lookup" call when renaming.  Now that
we don't undo a macro-alias on renaming anymore, but only look up
already-renamed identifiers in the syntax env, this case is no longer
triggered, so we fall back to the "else" clause, which will cause the
symbol to be aliased normally.

All the sample code snippets from that ticket behave identically with
and without this patch.

Signed-off-by: Evan Hanson <evhan@foldling.org>