Move some procedures from chicken.eval to chicken.platform
These are: chicken-home, extension-information, repository-path and
installation-repository.

provide and provided? should probably go to chicken.load, considering
those directly interact with require, which has more to do with code
loading than platform features. So they stay put in chicken.eval.

The entire module is moved down to the end of library.scm, because
some of these definitions take strings from C using the FFI, and the
FFI will inject calls to ##sys#make-c-string in the code, which is
defined pretty late in library.scm.

Signed-off-by: Evan Hanson <evhan@foldling.org>