Restore macro-expansion in canonicalize-body's "main" loop
In commit c9220247dbcdf6fd39697b428cfd40068244219a, we removed a little
bit too much: the original code's expansion in the main loop would
expand begins and defines at the root level, which means they'd be
"flattened" into the same letrec.

This broke a few situations that are technically somewhat iffy, like

  (let () (define (foo) bar) (begin (define (bar) 1) (foo)))

which would error out instead of returning 1, because the begin would
stop the main loop and inject a new letrec, acting as a barrier, which
means later definitions are in a new scope that the earlier definitions
would not be able to see.

Also, if there's an macro that expands to "define", that would also be
put in its own letrec because macro-expansion would stop the current
collection of vars and mvars, much like the aforementioned begin.

Signed-off-by: Evan Hanson <evhan@foldling.org>