Remove ##sys#nodups and ##sys#del
The ##sys#nodups and corresponding ##sys#del procedures are just
differently named (and specialisable) versions of delete-duplicates
and delete from SRFI-1.  So, we load mini-srfi-1.scm into csi.scm, and
get rid of those definition in library.scm.

We also get rid of the optional arguments in the SRFI-1 "delete" and
"delete-duplicates" definitions because that's completely unnecessary
for a fast internal API.

Signed-off-by: Evan Hanson <evhan@foldling.org>