Get rid of interactive prompts during install on mingw.
Due to dae44b93b02118c70c98cfd47041061b66bc0f0f, we now install
manpages to a different target filename than the source filename has.
Because we're using XCOPY, when we're installing to an empty
directory, it asks us whether we intend the target to be a
differently-named target *file*name or a newly-to-create *directory*
under which it should place the file with the original source's
filename.

Because XCOPY has no switch to tell it to force interpretation of the
final path component as a file, only as a directory (sigh,
Windows...), we can't use it.  Luckily, we only ever use the /Y flag,
which COPY also has.  So, COPY is effectively equivalent to XCOPY in
all the situations we use it in.  Therefore, switch to copy instead.

Signed-off-by: Evan Hanson <evhan@foldling.org>