Add syntax exports to chicken.condition
The macros condition-case and handle-exceptions really belong to
chicken.condition, but adding macros to library.scm is problematic
because the module and eval units rely on library, whereas to add
macros you'll need to use those two, causing a cyclic dependency.

This dependency cycle is why we have a separate chicken-syntax unit.
So, we keep the definitions there, but we add them to a separate
syntax environment which we use in a hand-rolled import library, which
we use in lieu of an emitted import library based on the module
definition (which does not contain said macros).

Because we no longer emit the import library, the compiler would
generate a call to eval to register the module at least in the current
compilation unit.  To suppress this, we add -no-module-registration
when compiling library.scm.

Finally, to allow compilation with an older version CHICKEN, we use
the original syntax environment from chicken if the new
chicken.condition-macro-environment environment is undefined.  This is
strictly incorrect because too many macros will be exported by the
chicken.condition module, but that's not a problem in practice, and
once we have a bootstrap CHICKEN we can get rid of this hack.

Signed-off-by: Evan Hanson <evhan@foldling.org>