Replace file-modification-time setter with set-file-times! procedure
This reduces several inconsistencies and any resulting confusion:

- file-access-time and file-change-time have no associated setter
- The setter for file-modification-time sets both mtime AND ctime
- The getters all accept strings, ports and file descriptors; the
   setter only accepts a string.

While at it, the new procedure also makes it possible to omit the
timestamps (in which case the current time is assumed), supply only
one (in which case the old behaviour stays: we set both timestamps to
the supplied time) or both (in which case you can set either to a
different value).  If #f is supplied, the specific time is unchanged.

This behaviour is maximally compatible with the "specify both or none"
behaviour from SCSH's "set-file-times" procedure (note the missing
bang though), and with MIT Scheme's "set-file-times!" procedure where
passing in #f means to avoid modifying the corresponding time.

Signed-off-by: Kooda <kooda@upyum.com>