Improve consistency of posix interface between Windows and Unix
Firstly, the export lists of the posix module didn't match up between
the two platforms, which meant any program using process-spawn, the
spawn flags or the open/noinherit flag for create-pipe (and for all
the low-level "open" procedures in general, in fact), would have to
cond-expand on platform.  Now we define them as no-ops for UNIX and
export them, so you can import them with impunity.  These extra flags
have also been added to types.db

Secondly, the types of create-pipe, process, process* and
process-execute didn't match up: the extra "exact-flag" argument for
the process procedures and the extra "mode" argument for create-pipe
were accepted only on Windows, and it would result in compilation
warnings due to types.db not knowing about this.  On UNIX, passing the
extra "mode" argument to create-pipe would result in an error, but due
to the specifics of how optional arguments are handled by CHICKEN,
passing the extra exact-flag argument to the process procedures on
UNIX wouldn't be a problem in practice.  Nevertheless, you'd still get
a compilation warning.  This has now been fixed by explicitly adding
the (unused) extra arguments to the procedures and the types.db
entries.

Finally, process-spawn was only defined on Windows, so any program
that used it couldn't even be compiled on UNIX.  Now we move
define-unimplemented to posix-common and define process-spawn as
unimplemented on UNIX.  An entry for process-spawn has also been added
to types.db.  Its runtime type check has been improved from checking
for any exact number (which is incorrect) to checking for a fixnum.

Signed-off-by: Kooda <kooda@upyum.com>