Rename change-file-mode to set-file-permissions! for consistency
This includes a SRFI-17 setter on file-permissions as well.  The new
setter is moved from (chicken file) to (chicken file posix), because
the getter lives in that module too.

file-permissions is changed to return just the permissions, so that
the values used by setter and getter are symmetric.

Before, the getter would also return the file type from stat(), as in
it returned the raw value of st_mode.  This makes no sense for the
name "file-permissions", I'd expect only the permissions.

Note that this doesn't remove any functionality: we can still get the
file type or any of the special bits using other getters, and if we
want we can still use "file-stat" to get the raw underlying values.

Finally, the setter is modified to match the getter in accepting both
a string and a fd or port, using fchmod.  On Windows, fchmod is
implemented using GetFinalPathNameByHandle() followed by chmod().  To
make this work, the Windows API version has been bumped to Vista.
It's probably better to officially require Windows 7 or newer, though.

Signed-off-by: Kooda <kooda@upyum.com>